+++ CHANGELOG.md
@@ -1,1 +1,18 @@
-2.5.1 - Fix desync booboo by not copying the table
+2.6.0:
+Blueprinter automatic planner expanded
+Now cycles through a list of blueprints and gets or creates them
+Lots of small fixes and improvments
+Fix Reversed locale for pickup/loot setting, Closes #21
+Add Position.Incremenet and Inventory.each_reverse.
+Switched Instant Blueprint to check collision box not selection.
+Update Quick UG Belt to use Position.Increment, make more effecient.
+Remove extra log spam from blueprint/balancers.
+Fix Quick BP off center.
+Fix Locale for inventory editor.
+Fix Balancer replacer not working on balancer bps only.
+Fix Dollies crash moving ghosts. Last user not being set in some cases.
+Create placeable off grid flying text for incrememnter.
+Fix last user checks in many spots
+Convert to player.mod_settings in many spots.
+Changelog entry
+Increment Version